requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
supabase==2.10.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
